refac(test): refactor setup for both case and case type access tests
- Refactor test setup for users and organizations into conftest.py (code deduplication, pytest will find this fixture at a higher level and inject it in every test in the directory and the subdirectories.
- Add new test type for edge cases in enum (for not sharing the same Env instance between tests with empty dictionary and filled dictionary when running from launch.json)
