+++ test/test_client/enum.py
@@ -7,6 +7,9 @@ class TestType(Enum):
     CASEDB_CUSTOM = "CASEDB_CUSTOM"
     CASEDB_INTEGRATION_BUILD_DB = "CASEDB_INTEGRATION_BUILD_DB"
     CASEDB_INTEGRATION_CASE_ACCESS = "CASEDB_INTEGRATION_CASE_ACCESS"
+    CASEDB_INTEGRATION_CASE_ACCESS_EDGE_CASES = (
+        "CASEDB_INTEGRATION_CASE_ACCESS_EDGE_CASES"
+    )
     CASEDB_INTEGRATION_CASE_VALIDATION = "CASEDB_INTEGRATION_CASE_VALIDATION"
     CASEDB_INTEGRATION_CONTENT = "CASEDB_INTEGRATION_CONTENT"
     CASEDB_INTEGRATION_REFDATA_ACCESS = "CASEDB_INTEGRATION_REFDATA_ACCESS"
